+++ block-padding/CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 (2020-12-08)
+### Changed
+- The `Padding` trait methods now work with blocks instead of byte slices. ([#113])
+
+[#113]: https://github.com/RustCrypto/utils/pull/113
+
 ## 0.2.1 (2020-08-14)
 ### Added
-- `Copy`, `Clone`, and `Debug` trait implementations for padding types ([#78])
+- `Copy`, `Clone`, and `Debug` trait implementations for padding types. ([#78])
 
 [#78]: https://github.com/RustCrypto/utils/pull/78